fix: fix pattern
